Rename package name of applycrds to make it useable
